requirements.txt
anthropic
gradio
requests
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-groq
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
supabase
arxiv
pymupdf
wikipedia
pgvector
python-dotenv